requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
groq==0.4.2
sentence-transformers==2.2.2
pgvector==0.2.4
PyPDF2==3.0.1
python-docx==0.8.11
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
